requirements_test.txt
coverage
coveralls
django-nose
pep8==1.5.7
flake8
mock
nose-exclude
nose-testconfig
selenium
tox
git+https://github.com/openstack/bandit.git   # PyPi version is out of date
